requirements.txt
google-genai
pypdf
python-dotenv
faiss-cpu
numpy
